Refactor form handling and Excel export
Refactored form handling for improve flexibility and improved Excel export functionality. In Modals/index.tsx, replaced 'selectedValue' prop with 'formBlueprint' to input the entire form blueprint object for better flexibility. In generateExcelSheet.ts, the function generateExcelSheet() is updated to use 'formBlueprint' which allows better handling of form responses and inclusion of an index as the first column in exported Excel sheet.

In FormsTable.tsx, code is reorganized for better readability and a new function exportForm() is added to enable exporting forms to Excel Sheets right from the forms table. This change streamlines workflow by providing a direct way to export form responses. Now buttons for 'send', 'export' and 'preview' actions are available for each form in the table.