Add support for individual email input in modal
This commit aims to enhance email input functionality in the modal component. Previously, users were only able to select from predefined options. Changes in this commit allow the input of individual emails, which are added to the email list upon the input field losing focus. This simplifies email input for users and makes the user interface more intuitive.