Refactor form expiration check; support formBlueprint data
Moved isFormExpired function to a new utility file, formUtils, as it was being used in multiple components and needed to be more easily accessible. Removed method from [id].tsx and imported from formUtils instead.

Modified Modals and FormsTable component to account for change in data structure within formBlueprint resulting from the updated Firestore query. Added Tooltip in Modal to indicate when form is expired and disabled share button accordingly for a better user experience.

The refactoring was necessary to avoid code duplication and, the tooltip was added to enhance user experience by giving feedback about the form's state.