Simplify README.md by removing unnecessary details
The README.md file was simplified by removing a large chunk of text and images which weren't crucial for understanding the project steps. These removed parts included some marketing information such as the logo, feature badges, product images, and promotional texts.

This was done to make it easier for users to quickly grasp the essential information on how to set up and use the project. The Quick Start section now stands out, making it more accessible and straightforward for users. The changes contribute to a cleaner, more efficient document.